requirements.txt
aiohttp
beautifulsoup4
PyYAML
